working example of solution of the task as given below:
Create a script, excel formula or stored procedure to compute the number
of blocks of each load testing block sizes

for any given number 

of load testing hours that a client can buy from Cyara. 

The script/excel formula/stored procedure must take into account the
least cost to the client assuming 

* 4 hour block costs $100, 6 hour block costs $150, 9 hour costs $225,
20 hour block costs $480

In the script we take for assumption that more bigger blocks would give
less price eventually if compared to price obtained from the number
represented by smaller blocks.
It means script doesn't check ALL possible variants of calculation of
the given load testing hours with help of basic block but rather trys to
figure the less amount of blocks these hours could be assembled with.

As such, 24 could be 4*6, 6*4, 9+9+6 and 20+4
it would give us 600$ for all but 20+4 ($580)